refactor: lot 6 — déplacer fiscalitySettings.ts dans utils/cache/ et supprimer src/types/
- Déplacer src/types/fiscalitySettings.ts → src/utils/cache/fiscalitySettings.ts (2 consommateurs locaux)
- Mettre à jour imports dans fiscalSettingsCache.ts et fiscalitySettingsMigrator.ts
- Supprimer le répertoire src/types/ (vide)
- Vérification : typecheck ✅
